Crate rumi_http

rumi-http: HTTP domain for request matching

This crate provides two layers:

User API: Gateway API HttpRouteMatch for configuration

Data Plane API: HttpMessage (indexed from ProcessingRequest) for runtime

Architecture

Gateway API HttpRouteMatch (config)
        ↓ compile()
rumi Matcher<HttpMessage, A>
        ↓ evaluate()
HttpMessage (indexed from ext_proc ProcessingRequest)

Example

use rumi_http::prelude::*;

// Config time: compile Gateway API match to rumi matcher
let route_match = HttpRouteMatch {
    path: Some(HttpPathMatch::PathPrefix { value: "/api".into() }),
    ..Default::default()
};
let matcher = route_match.compile::<&str>("api_backend");

// Runtime: index ProcessingRequest into HttpMessage, then evaluate
let msg = HttpMessage::from(&processing_request);
let result = matcher.evaluate(&msg);

Modules

prelude

Prelude for convenient imports.

Structs

AuthorityInput

Extracts the :authority pseudo-header from HttpMessage.

HeaderInput

Extracts a header value from HttpMessage.

HttpMessage

Indexed view over HTTP request data for efficient matching.

HttpRequest

Simple HTTP request context for matching.

HttpRequestBuilder

Builder for HttpRequest.

HttpRouteMatch

HTTPRouteMatch defines the predicate used to match requests to a given action. Multiple match types are ANDed together, i.e. the match will evaluate to true only if all conditions are satisfied.

MethodInput

Extracts the HTTP method from HttpMessage.

PathInput

Extracts the request path (without query string) from HttpMessage.

ProcessingRequest

This represents the different types of messages that the data plane can send to an external processing server. [#next-free-field: 12]

ProcessingResponse

This represents the different types of messages the server may send back to the data plane when the observability_mode field in the received ProcessingRequest is set to false.

QueryParamInput

Extracts a query parameter value from HttpMessage.

SchemeInput

Extracts the :scheme pseudo-header from HttpMessage.

SimpleHeaderInput

Extracts a header from simple HttpRequest.

SimpleMethodInput

Extracts the HTTP method from simple HttpRequest.

SimplePathInput

Extracts the path from simple HttpRequest.

SimpleQueryParamInput

Extracts a query parameter from simple HttpRequest.

Enums

HttpHeaderMatch

HTTPHeaderMatch describes how to select a HTTP route by matching HTTP request headers.

HttpPathMatch

HTTPPathMatch describes how to select a HTTP route by matching the HTTP request path.

HttpQueryParamMatch

HTTPQueryParamMatch describes how to select a HTTP route by matching HTTP query parameters.

Traits

HttpRouteMatchExt

Extension trait for compiling HttpRouteMatch to rumi Matcher.

Functions

compile_route_matches

Compile multiple HttpRouteMatch entries into a single Matcher.

Type Aliases

HttpMethod

HTTPMethod describes how to select a HTTP route by matching the HTTP method as defined by RFC 7231 and RFC 5789. The value is expected in upper case.