rumi_http

Module prelude

Prelude for convenient imports.

Re-exports

pub use super::AuthorityInput;

pub use super::HeaderInput;

pub use super::HttpMessage;

pub use super::HttpRequest;

pub use super::HttpRequestBuilder;

pub use super::HttpRouteMatchExt;

pub use super::MethodInput;

pub use super::PathInput;

pub use super::QueryParamInput;

pub use super::SchemeInput;

pub use super::SimpleHeaderInput;

pub use super::SimpleMethodInput;

pub use super::SimplePathInput;

pub use super::SimpleQueryParamInput;

Structs

BoolMatcher

Boolean equality matcher.

ContainsMatcher

Contains string matcher.

ExactMatcher

Exact string equality matcher.

FieldMatcher

A field matcher: predicate + on_match.

HttpRouteMatch

HTTPRouteMatch defines the predicate used to match requests to a given action. Multiple match types are ANDed together, i.e. the match will evaluate to true only if all conditions are satisfied.

Matcher

Top-level matcher with first-match-wins semantics.

PrefixMatcher

Prefix string matcher.

ProcessingRequest

This represents the different types of messages that the data plane can send to an external processing server. [#next-free-field: 12]

ProcessingResponse

This represents the different types of messages the server may send back to the data plane when the observability_mode field in the received ProcessingRequest is set to false.

RadixTree

A radix tree (compressed trie) for efficient prefix lookups.

SinglePredicate

A single predicate: combines a DataInput with an InputMatcher.

SuffixMatcher

Suffix string matcher.

Enums

HttpHeaderMatch

HTTPHeaderMatch describes how to select a HTTP route by matching HTTP request headers.

HttpPathMatch

HTTPPathMatch describes how to select a HTTP route by matching the HTTP request path.

HttpQueryParamMatch

HTTPQueryParamMatch describes how to select a HTTP route by matching HTTP query parameters.

MatcherError

Errors from matcher validation.

MatcherTree

Tree-based matcher using map lookups instead of predicate evaluation.

MatchingData

The erased data type that flows between DataInput and InputMatcher.

OnMatch

Defines what to do when a predicate matches.

Predicate

Composite predicate with boolean logic.

StringMatcher

Unified string matcher matching xDS StringMatcher proto.

Traits

CustomMatchData

Extension trait for custom matching data types.

DataInput

Extracts data from a domain-specific context.

InputMatcher

Matches against type-erased MatchingData.

Type Aliases

HttpMethod

HTTPMethod describes how to select a HTTP route by matching the HTTP method as defined by RFC 7231 and RFC 5789. The value is expected in upper case.