rumi_http::prelude

Struct HttpRouteMatch

pub struct HttpRouteMatch {
    pub path: Option<HttpPathMatch>,
    pub headers: Option<Vec<HttpHeaderMatch>>,
    pub query_params: Option<Vec<HttpQueryParamMatch>>,
    pub method: Option<String>,
}

HTTPRouteMatch defines the predicate used to match requests to a given action. Multiple match types are ANDed together, i.e. the match will evaluate to true only if all conditions are satisfied.

For example, the match below will match a HTTP request only if its path starts with /foo AND it contains the version: v1 header:

match:
  path:
    value: "/foo"
  headers:
  - name: "version"
    value "v1"

Fields
path: Option<HttpPathMatch>

Path specifies a HTTP request path matcher. If this field is not specified, a default prefix match on the โ/โ path is provided.

headers: Option<Vec<HttpHeaderMatch>>

Headers specifies HTTP request header matchers. Multiple match values are ANDed together, meaning, a request must match all the specified headers to select the route.

query_params: Option<Vec<HttpQueryParamMatch>>

QueryParams specifies HTTP query parameter matchers. Multiple match values are ANDed together, meaning, a request must match all the specified query parameters to select the route.

method: Option<String>

Method specifies HTTP method matcher.

When specified, this route will be matched only if the request has the specified method.

Support: Extended