rumi_http

Struct ProcessingRequest

pub struct ProcessingRequest {
    pub metadata_context: Option<Metadata>,
    pub attributes: HashMap<String, Struct>,
    pub observability_mode: bool,
    pub protocol_config: Option<ProtocolConfiguration>,
    pub request: Option<Request>,
}

This represents the different types of messages that the data plane can send to an external processing server. [#next-free-field: 12]

Fields
metadata_context: Option<Metadata>

Dynamic metadata associated with the request.

attributes: HashMap<String, Struct>

The values of properties selected by the request_attributes or response_attributes list in the configuration. Each entry in the list is populated from the standard : ref:attributes <arch_overview_attributes> supported in the data plane.

observability_mode: bool

Specify whether the filter that sent this request is running in :ref:observability_mode <envoy_v3_api_field_extensions.filters.http.ext_proc.v3.ExternalProcessor.observability_mode> and defaults to false.

A value of false indicates that the server must respond to this message by either sending back a matching ProcessingResponse message, or by closing the stream.

A value of true indicates that the server should not respond to this message, as any responses will be ignored. However, it may still close the stream to indicate that no more messages are needed.

protocol_config: Option<ProtocolConfiguration>

Specify the filter protocol configurations to be sent to the server. protocol_config is only encoded in the first ProcessingRequest message from the client to the server.

request: Option<Request>

Each request message will include one of the following sub-messages. Which ones are set for a particular HTTP request/response depend on the processing mode.