rumi_http

Trait ProcessingRequestExt

pub trait ProcessingRequestExt {
    // Required methods
    fn request_headers(&self) -> Option<&HttpHeaders>;
    fn response_headers(&self) -> Option<&HttpHeaders>;
    fn get_request_header(&self, name: &str) -> Option<&str>;
    fn get_path(&self) -> Option<&str>;
    fn get_method(&self) -> Option<&str>;
    fn get_authority(&self) -> Option<&str>;
    fn get_scheme(&self) -> Option<&str>;
}

Extension trait for ProcessingRequest to extract HTTP data.

Required Methods

fn request_headers(&self) -> Option<&HttpHeaders>

Get the request headers if this is a request_headers message.

fn response_headers(&self) -> Option<&HttpHeaders>

Get the response headers if this is a response_headers message.

fn get_request_header(&self, name: &str) -> Option<&str>

Get a header value by name from request headers. Header names are matched case-insensitively (all stored lowercase).

fn get_path(&self) -> Option<&str>

Get the :path pseudo-header from request headers.

fn get_method(&self) -> Option<&str>

Get the :method pseudo-header from request headers.

fn get_authority(&self) -> Option<&str>

Get the :authority pseudo-header from request headers.

fn get_scheme(&self) -> Option<&str>

Get the :scheme pseudo-header from request headers.